// apps/mobile/src/app/home/models/filter-type.enum.ts
export enum FilterType {
  ACCOMODATION = 'ACCOMODATION',
  ATRACTIONS = 'ATRACTIONS',
  RESTAURANTS = 'RESTAURANTS',
  TRANSPORT = 'TRANSPORT',
  LOOKOUTS = 'LOOKOUTS',
  CAFES = 'CAFES',
  RECREATION = 'RECREATION',
  CULTURE = 'CULTURE',
  PROTECTED_NATURAL_ASSETS = 'PROTECTED_NATURAL_ASSETS',
  EVENTS = 'EVENTS',
}